requirements.txt
google-generativeai
stremlit
PIL